Changes:  - workaround for X DISPLAY variable missing problem  - resolved some compiler warnings (gcc 4 on Linux)
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5310 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/utilities/source/font_renderer.cpp
@@ -301,7 +301,7 @@ namespace mcrl2 {
     {
       // Wrapped text to fit into bounding box. 
       
-      const size_t addpos = (xRgt-xLft)/(CHARWIDTH*scale);
+      const size_t addpos = static_cast< size_t >((xRgt-xLft)/(CHARWIDTH*scale));
       double transx = 0;
       double transy = 0;
       std::string subs;
